backend/requirements.txt
flask
python-dotenv
google-generativeai
PyPDF2
docx2txt
